resolving comments and suggestions (part1)
