--- espnet/nets/chainer_backend/asr_transformer/__init__.py
@@ -1,4 +1,0 @@
-from espnet.nets.chainer_backend.asr_transformer.e2e import E2E  # NOQA
-from espnet.nets.chainer_backend.asr_transformer.e2e import VaswaniRule  # NOQA
-from espnet.nets.chainer_backend.asr_transformer.net_utils import add_arguments  # NOQA
-from espnet.nets.chainer_backend.asr_transformer.net_utils import PlotAttentionReport  # NOQA
