+++ espnet/nets/chainer_backend/e2e_asr_transformer.py
@@ -6,9 +6,11 @@ from chainer import reporter
 import chainer.functions as F
 from chainer.training import extension
 
-from espnet.nets.chainer_backend.asr_transformer.attention import MultiHeadAttention
-from espnet.nets.chainer_backend.asr_transformer.decoders import Decoder
-from espnet.nets.chainer_backend.asr_transformer.encoders import Encoder
+from distutils.util import strtobool
+
+from espnet.nets.chainer_backend.attentions_transformer import MultiHeadAttention
+from espnet.nets.chainer_backend.decoders_transformer import Decoder
+from espnet.nets.chainer_backend.encoders_transformer import Encoder
 
 import logging
 import math
@@ -16,6 +18,24 @@ import numpy as np
 
 MAX_DECODER_OUTPUT = 5
 MIN_VALUE = float(np.finfo(np.float32).min)
+
+
+def add_arguments(parser):
+    group = parser.add_argument_group("transformer model setting")
+    group.add_argument("--transformer-init", type=str, default="pytorch",
+                       help='how to initialize transformer parameters')
+    group.add_argument("--transformer-input-layer", type=str, default="conv2d",
+                       choices=["conv2d", "linear", "embed"],
+                       help='transformer input layer type')
+    group.add_argument('--transformer-attn-dropout-rate', default=None, type=float,
+                       help='dropout in transformer attention. use --dropout-rate if None is set')
+    group.add_argument('--transformer-lr', default=10.0, type=float,
+                       help='Initial value of learning rate')
+    group.add_argument('--transformer-warmup-steps', default=25000, type=int,
+                       help='optimizer warmup steps')
+    group.add_argument('--transformer-length-normalized-loss', default=True, type=strtobool,
+                       help='normalize loss by length')
+    return parser
 
 
 class VaswaniRule(extension.Extension):
@@ -279,9 +299,7 @@ class E2E(chainer.Chain):
             xs, x_mask, ilens = self.encoder(x_block[None, :, :], ilens)
             logging.info('Encoder size: ' + str(xs.shape))
             if recog_args.ctc_weight > 0.0:
-                lpz = self.ctc.log_softmax(xs)[0]  # NOQA
-            else:
-                lpz = None  # NOQA
+                raise NotImplementedError('use joint ctc/tranformer decoding. WIP')
             if recog_args.beam_size == 1:
                 logging.info('Use greedy search implementation')
                 ys = xp.full((1, 1), self.sos)
@@ -292,7 +310,6 @@ class E2E(chainer.Chain):
                     yy_mask *= self.make_history_mask(ys)
                     xy_mask = self.make_attention_mask(ys, xp.array(x_mask))
                     out = self.decoder(ys, yy_mask, xs, xy_mask).reshape(batch, -1, self.odim)
-                    # out = self.output(out, n_batch_axes=2)
                     prob = F.log_softmax(out[:, -1], axis=-1)
                     max_prob = prob.array.max(axis=1)
                     next_id = F.argmax(prob, axis=1).array.astype(np.int64)
